+++ test/reduction_tree_generator_test.dart
@@ -82,7 +82,7 @@ void main() {
     final vec = <Logic>[];
     // First sum will be length *(length-1) /2
     for (var i = 0; i < length; i++) {
-      vec.add(Logic(width: width)..put(i));
+      vec.add(Logic(width: width)..inject(i));
     }
     const radix = 4;
     final prefixAdd = ReductionTreeGenerator(
@@ -162,7 +162,7 @@ void main() {
     final vec = <Logic>[];
     // First sum will be length *(length-1) /2
     for (var i = 0; i < length; i++) {
-      vec.add(Logic(width: width)..put(i));
+      vec.add(Logic(width: width)..inject(i));
     }
     const reduce = 4;
     final prefixAdd = ReductionTreeGenerator(